Move sidebar tab-title effect to single owner
